+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `VQT Platform GameObject Remover` component to remove game objects depending on the build platform (NDMF is required).
+
 ### Fixed
 - Fixed an issue where the converter fails when the avatar object name starts or ends with whitespaces.